+++ poutyne/framework/experiment.py
@@ -881,55 +881,7 @@ class Experiment:
 
         return ret
 
-    def predict(self, x, **kwargs) -> Any:
-        """
-        Returns the predictions of the network given a dataset ``x``, where the tensors are
-        converted into Numpy arrays.
-
-        Args:
-            x (Union[~torch.Tensor, ~numpy.ndarray] or Union[tuple, list] of Union[~torch.Tensor, ~numpy.ndarray]):
-                Input to the model. Union[Tensor, ndarray] if the model has a single input.
-                Union[tuple, list] of Union[Tensor, ndarray] if the model has multiple inputs.
-            checkpoint (Union[str, int]): Which model checkpoint weights to load for the prediction.
-
-                - If 'best', will load the best weights according to ``monitor_metric`` and ``monitor_mode``.
-                - If 'last', will load the last model checkpoint.
-                - If int, will load the checkpoint of the specified epoch.
-                - If a path (str), will load the model pickled state_dict weights (for instance, saved as
-                  ``torch.save(a_pytorch_network.state_dict(), "./a_path.p")``).
-
-                This argument has no effect when logging is disabled. (Default value = 'best')
-            kwargs: Any keyword arguments to pass to :func:`~Model.predict()`.
-
-        Returns:
-            Return the predictions in the format outputted by the model.
-        """
-        return self._predict(self.model.predict, x, **kwargs)
-
-    def predict_dataset(self, dataset, **kwargs) -> Any:
-        """
-        Returns the predictions of the network given a dataset, where the tensors are
-        converted into Numpy arrays.
-
-        Args:
-            dataset (~torch.utils.data.Dataset): Dataset. Must not return ``y``, just ``x``.
-            checkpoint (Union[str, int]): Which model checkpoint weights to load for the prediction.
-
-                - If 'best', will load the best weights according to ``monitor_metric`` and ``monitor_mode``.
-                - If 'last', will load the last model checkpoint.
-                - If int, will load the checkpoint of the specified epoch.
-                - If a path (str), will load the model pickled state_dict weights (for instance, saved as
-                  ``torch.save(a_pytorch_network.state_dict(), "./a_path.p")``).
-
-                This argument has no effect when logging is disabled. (Default value = 'best')
-            kwargs: Any keyword arguments to pass to :func:`~Model.predict_dataset()`.
-
-        Returns:
-            Return the predictions in the format outputted by the model.
-        """
-        return self._predict(self.model.predict_dataset, dataset, **kwargs)
-
-    def predict_generator(self, generator, **kwargs) -> Any:
+    def infer(self, generator, **kwargs) -> Any:
         """
         Returns the predictions of the network given batches of samples ``x``, where the tensors are
         converted into Numpy arrays.
@@ -957,6 +909,54 @@ class Experiment:
             for the batches is returned with tensors converted into Numpy arrays.
         """
         return self._predict(self.model.predict_generator, generator, **kwargs)
+
+    def infer_dataset(self, dataset, **kwargs) -> Any:
+        """
+        Returns the inferred predictions of the network given a dataset, where the tensors are
+        converted into Numpy arrays.
+
+        Args:
+            dataset (~torch.utils.data.Dataset): Dataset. Must not return ``y``, just ``x``.
+            checkpoint (Union[str, int]): Which model checkpoint weights to load for the prediction.
+
+                - If 'best', will load the best weights according to ``monitor_metric`` and ``monitor_mode``.
+                - If 'last', will load the last model checkpoint.
+                - If int, will load the checkpoint of the specified epoch.
+                - If a path (str), will load the model pickled state_dict weights (for instance, saved as
+                  ``torch.save(a_pytorch_network.state_dict(), "./a_path.p")``).
+
+                This argument has no effect when logging is disabled. (Default value = 'best')
+            kwargs: Any keyword arguments to pass to :func:`~Model.predict_dataset()`.
+
+        Returns:
+            Return the predictions in the format outputted by the model.
+        """
+        return self._predict(self.model.predict_dataset, dataset, **kwargs)
+
+    def infer_data(self, x, **kwargs) -> Any:
+        """
+        Returns the inferred predictions of the network given a dataset ``x``, where the tensors are
+        converted into Numpy arrays.
+
+        Args:
+            x (Union[~torch.Tensor, ~numpy.ndarray] or Union[tuple, list] of Union[~torch.Tensor, ~numpy.ndarray]):
+                Input to the model. Union[Tensor, ndarray] if the model has a single input.
+                Union[tuple, list] of Union[Tensor, ndarray] if the model has multiple inputs.
+            checkpoint (Union[str, int]): Which model checkpoint weights to load for the prediction.
+
+                - If 'best', will load the best weights according to ``monitor_metric`` and ``monitor_mode``.
+                - If 'last', will load the last model checkpoint.
+                - If int, will load the checkpoint of the specified epoch.
+                - If a path (str), will load the model pickled state_dict weights (for instance, saved as
+                  ``torch.save(a_pytorch_network.state_dict(), "./a_path.p")``).
+
+                This argument has no effect when logging is disabled. (Default value = 'best')
+            kwargs: Any keyword arguments to pass to :func:`~Model.predict()`.
+
+        Returns:
+            Return the predictions in the format outputted by the model.
+        """
+        return self._predict(self.model.predict, x, **kwargs)
 
     def _predict(self,
                  evaluate_func: Callable,